+++ src/maths/sqrt.c
@@ -5,6 +5,7 @@
 ** square root function
 */
 
+#include <stddef.h>
 #include "tlcmaths.h"
 
 static const int accurency = 100;
@@ -27,7 +28,7 @@ static long double x_sqrtl_find(long double x, long double lower,
 }
 
 static void x_sqrtl_bouds(long double *lower, long double *higher,
-    long double (*power)(long double x, long double y), long double x)
+    long double (*power)(long double x, size_t y), long double x)
 {
     long double tmp_lo = 0;
     long double tmp_hi = 0;
@@ -38,37 +39,37 @@ static void x_sqrtl_bouds(long double *lower, long double *higher,
         tmp_lo *= 100;
     }
     tmp_hi = power(*higher, 2);
-    while (0.01 * tmp_hi > x) {
-        *higher *= 0.1;
-        tmp_hi *= 0.01;
+    while (((long double) 0.01) * tmp_hi > x) {
+        *higher *= (long double) 0.1;
+        tmp_hi *= (long double) 0.01;
     }
 }
 
 double x_sqrt(double x)
 {
-    long double lower = x_mind(1, x);
-    long double higher = x_maxd(1, x);
+    long double lower = (long double) x_mind(1, x);
+    long double higher = (long double) x_maxd(1, x);
     double res = 0;
 
     if (x <= 1) {
         return (x);
     }
-    x_sqrtl_bouds(&lower, &higher, x_powl, x);
-    res = x_sqrtl_find(x, lower, higher, accurency);
+    x_sqrtl_bouds(&lower, &higher, x_powl, (long double) x);
+    res = (double) x_sqrtl_find((long double) x, lower, higher, accurency);
     return (res);
 }
 
 float x_sqrtf(float x)
 {
-    long double lower = x_minf(1, x);
-    long double higher = x_maxf(1, x);
+    long double lower = (long double) x_minf(1, x);
+    long double higher = (long double) x_maxf(1, x);
     float res = 0;
 
     if (x <= 1) {
         return (x);
     }
-    x_sqrtl_bouds(&lower, &higher, x_powl, x);
-    res = x_sqrtl_find(x, lower, higher, accurency * 10);
+    x_sqrtl_bouds(&lower, &higher, x_powl, (long double) x);
+    res = (float) x_sqrtl_find((long double) x, lower, higher, accurency * 10);
     return (res);
 }
