+++ src/stdlibs/calloc.c
@@ -5,9 +5,11 @@
 ** malloc n char and fill with 0
 */
 
+#include <stddef.h>
 #include <stdlib.h>
+#include "tlcstdlibs.h"
 
-char *x_calloc(int n)
+char *x_calloc(size_t n)
 {
     char *new = NULL;
 
@@ -15,7 +17,7 @@ char *x_calloc(int n)
     if (new == NULL) {
         return (NULL);
     }
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         new[i] = '\0';
     }
     return (new);
