+++ src/strings/strrev.c
@@ -11,13 +11,13 @@
 char *x_strrev(char *str)
 {
     char c = '\0';
-    int len = 0;
+    size_t len = 0;
 
     if (str == NULL) {
         return (NULL);
     }
     len = x_strlen(str) - 1;
-    for (int i = 0; i <= len / 2; i++) {
+    for (size_t i = 0; i <= len / 2; i++) {
         c = str[i];
         str[i] = str[len - i];
         str[len - i] = c;
